refac(calculated): Align concepts of `ColumnNames` and `CalculatedNames`
