+++ source/geh_calculated_measurements/src/geh_calculated_measurements/opengeh_capacity_settlement/domain/calculated_names.py
@@ -1,0 +1,5 @@
+class CalculatedNames:
+    """Names of calculated columns."""
+
+    period_from_date = "period_from_date"
+    period_to_date = "period_to_date"
