+++ source/geh_calculated_measurements/src/geh_calculated_measurements/opengeh_capacity_settlement/domain/column_names.py
@@ -1,0 +1,12 @@
+class ColumNames:
+    # Input/output column names from contracts
+    date = "date"
+    child_metering_point_id = "child_metering_point_id"
+    metering_point_id = "metering_point_id"
+    observation_time = "observation_time"
+    quantity = "quantity"
+    child_period_from_date = "child_period_from_date"
+    child_period_to_date = "child_period_to_date"
+    # Ephemeral columns
+    selection_period_start = "selection_period_start"
+    selection_period_end = "selection_period_end"
